Create model modified subscription mechanism for pipelines
* Make model name constant
This is to enable logging during making subscriptions
* Add SubscriptionManager class
* Use notification mechanism when changing model/version
* Make pipeline creation subscribe to all used models/versions
* Prepare function for unsubscribing by pipeline definition

Test will be created in future MR's (integration tests).

JIRA:CVS-41768,CVS-41771
