+++ src/BUILD
@@ -43,6 +43,8 @@ cc_library(
         "model.hpp",
         "model_version_policy.cpp",
         "model_version_policy.hpp",
+        "modelchangesubscription.cpp",
+        "modelchangesubscription.hpp",
         "modelconfig.cpp",
         "modelconfig.hpp",
         "modelmanager.cpp",
@@ -63,6 +65,8 @@ cc_library(
         "ov_utils.hpp",
         "pipeline.cpp",
         "pipeline.hpp",
+        "pipelinedefinition.cpp",
+        "pipelinedefinition.hpp",
         "pipeline_factory.cpp",
         "pipeline_factory.hpp",
         "prediction_service.cpp",
